requirements.txt
# Core dependencies
elasticsearch>=8.15.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
openai>=1.12.0

# PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Image processing
Pillow>=10.0.0

# HTTP requests
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0

